% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_APW.R
\name{run_APW}
\alias{run_APW}
\title{Run AuPairWise}
\usage{
run_APW(
  exprs,
  out,
  stoich.pairs,
  n.factors = c(0, 1, 5, 10, 20, 50, 100),
  n.repeats = 3,
  dist = "other",
  mode = "post",
  ranked = FALSE,
  labels.default = "Stoichiometric pairs"
)
}
\arguments{
\item{out}{path for summary to be outputted}

\item{n.factors}{increment of noise factors}

\item{n.repeats}{number of repeats}

\item{dist}{dist}

\item{mode}{mode}

\item{ranked}{boolean of whether data is ranked}

\item{expr}{expression matrix the be assessed}

\item{stoic.pairs}{co-expressed genes used to calculate AUROC}

\item{label.default}{label for the stoic pairs}
}
\value{
summary of results
}
\description{
This function runs AuPairWise.
It calculates the average AUROC for the expression matrix
with different noise factors and outputs results into an Rdata file.
The noise factors and the number of repeats can be changed in the function parameters.
}
